Using spring profiles to set configuration
The spring profile in this project is application-test.yml. it resides in the same folder as default profile file application.yml
The spring profile naming format is
application-<profile name>.extn
An configuration parameter spring.profile.active=test is set in the application.yml, so that it overrides the values of the parameters.
FOr multiple environments, multiple yml files can be created and corresponding profile name can be passed as command line argument while execution
ex: java -jar project.jar --spring.profiles.active=prod, for this to work there should be a config file with name application-prod.yml